requirements.txt
fastapi==0.78.0
keras==2.9.0
Keras_Preprocessing==1.1.2
numpy==1.23.1
pandas==1.4.3
pydantic==1.9.1
python_twitter==3.5
scikit_learn==1.1.1
tensorflow-cpu==2.9.1
uvicorn
rq